refactor: move utils.js
