+++ src/commands/fs/move.js
@@ -1,5 +1,5 @@
 import fsPromises from 'fs/promises';
-import { copyFile } from './utils.js';
+import { copyFile } from '../utils.js';
 import Command from '../command.js';
 
 async function move() {
